Fix api error by adding changing id to _id find user correctly
